Add link icon to SwiftUI links
The formatting doesn't match the UIKit links, but I don't think that's a big problem.
The UIKit icons are on the left for simplicity so I don't have to create a custom UITableViewCell subclass while the SwiftUI icons are on the right so that the label can easily transition to the text field.